Remove some queries related to user guild
Changing design philospohy. Rather than selecting the server to see on
the client side, the website will just include a link to add bot to
the server. From the server you will be able to use slash commands
to create a queue in a given channel and then use view the queue at the
link the bot provides.